update players.json with nba dot com data
i wrote a quick script that grabbed the nba dot com players data, and added it
to players dot json. it now has the nba dot com player ids and team ids. i think
it also added some players in.

future work: it would be nice to have a script that checks teams and updates
them.